added getnate_dataset script.
